Refactor search functionality by extracting SummonerHeader from SummonerPage
- Create new SummonerHeader component containing search input, button and rivalry cart
- Improve code clarity by removing unnecessary state and handlers from SummonerPage
